# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LootBot multi-tenant store management API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface CsrfResponse {
  token: string;
}

export interface RegisterInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  OWNER: 'OWNER',
  SUPERADMIN: 'SUPERADMIN',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  csrfToken: string;
}

export interface StoreInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency: string;
}

export interface StoreUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency?: string;
}

export type StoreBotStatus = typeof StoreBotStatus[keyof typeof StoreBotStatus];


export const StoreBotStatus = {
  disconnected: 'disconnected',
  connected: 'connected',
  error: 'error',
} as const;

export interface Store {
  id: string;
  ownerId: string;
  name: string;
  slug: string;
  currency: string;
  botStatus: StoreBotStatus;
  createdAt: string;
}

export interface CategoryInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @maxLength 500 */
  description?: string;
}

export interface CategoryUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /** @maxLength 500 */
  description?: string;
}

export interface Category {
  id: string;
  storeId: string;
  name: string;
  /** @nullable */
  description: string | null;
  productCount: number;
  createdAt: string;
}

export interface CategoryPage {
  items: Category[];
  page: number;
  pageSize: number;
  total: number;
}

export interface ProductInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /** @maxLength 5000 */
  description: string;
  /**
     * @minimum 0
     * @maximum 999999999
     */
  price: number;
  /**
     * @minimum 0
     * @maximum 100000000
     */
  stock: number;
  /** @nullable */
  categoryId?: string | null;
  /**
     * @maxLength 80
     * @nullable
     */
  sku?: string | null;
  /**
     * @maxLength 2000
     * @nullable
     */
  imageUrl?: string | null;
  isPublished: boolean;
}

export interface ProductUpdate {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name?: string;
  /** @maxLength 5000 */
  description?: string;
  /**
     * @minimum 0
     * @maximum 999999999
     */
  price?: number;
  /**
     * @minimum 0
     * @maximum 100000000
     */
  stock?: number;
  /** @nullable */
  categoryId?: string | null;
  /**
     * @maxLength 80
     * @nullable
     */
  sku?: string | null;
  /**
     * @maxLength 2000
     * @nullable
     */
  imageUrl?: string | null;
  isPublished?: boolean;
}

export interface Product {
  id: string;
  storeId: string;
  /** @nullable */
  categoryId: string | null;
  /** @nullable */
  categoryName: string | null;
  /** @nullable */
  sku: string | null;
  name: string;
  description: string;
  price: number;
  stock: number;
  /** @nullable */
  imageUrl: string | null;
  isPublished: boolean;
  createdAt: string;
}

export interface ProductPage {
  items: Product[];
  page: number;
  pageSize: number;
  total: number;
}

export interface BotTokenInput {
  /**
     * @minLength 20
     * @maxLength 256
     */
  token: string;
}

export type BotStatusStatus = typeof BotStatusStatus[keyof typeof BotStatusStatus];


export const BotStatusStatus = {
  disconnected: 'disconnected',
  connected: 'connected',
  error: 'error',
} as const;

export interface BotStatus {
  connected: boolean;
  /** @nullable */
  username: string | null;
  /** @nullable */
  firstName: string | null;
  status: BotStatusStatus;
  /** @nullable */
  lastError: string | null;
}

export type DashboardSummaryBotStatus = typeof DashboardSummaryBotStatus[keyof typeof DashboardSummaryBotStatus];


export const DashboardSummaryBotStatus = {
  disconnected: 'disconnected',
  connected: 'connected',
  error: 'error',
} as const;

export interface DashboardSummary {
  storeId: string;
  productCount: number;
  categoryCount: number;
  publishedProductCount: number;
  orderCount: number;
  revenue: number;
  botStatus: DashboardSummaryBotStatus;
}

export interface ActivityEvent {
  id: string;
  action: string;
  summary: string;
  createdAt: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Sign in required
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Resource conflicts with existing data
 */
export type ConflictResponse = ErrorResponse;

/**
 * Too many attempts
 */
export type RateLimitedResponse = ErrorResponse;

/**
 * Telegram could not be reached or rejected the token
 */
export type UpstreamErrorResponse = ErrorResponse;

export type PageParameter = number;

export type PageSizeParameter = typeof PageSizeParameter[keyof typeof PageSizeParameter];


export const PageSizeParameter = {
  NUMBER_10: 10,
  NUMBER_25: 25,
  NUMBER_50: 50,
  NUMBER_100: 100,
} as const;

export type ListCategoriesParams = {
storeId: string;
/**
 * @minimum 1
 */
page?: PageParameter;
pageSize?: PageSizeParameter;
};

export type ListProductsParams = {
storeId: string;
/**
 * @minimum 1
 */
page?: PageParameter;
pageSize?: PageSizeParameter;
/**
 * @maxLength 120
 */
search?: string;
categoryId?: string;
};

export type GetDashboardSummaryParams = {
storeId: string;
};

export type GetDashboardActivityParams = {
storeId: string;
/**
 * @minimum 1
 * @maximum 25
 */
limit?: number;
};