# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LootBot multi-tenant store management API
 * OpenAPI spec version: 0.2.0
 */

export interface SuccessResponse {
  success: boolean;
}